refactor: tidy up the place
