+++ package.json
@@ -4,7 +4,7 @@
   "description": "libp2p WebRTC transport that includes a discovery mechanism provided by the signalling-star",
   "main": "src/index.js",
   "bin": {
-    "star-sig": "src/signalling/bin.js"
+    "star-sig": "src/sig-server/bin.js"
   },
   "browser": {
     "wrtc": false
