--- test/node.js
@@ -1,9 +1,0 @@
-'use strict'
-
-require('./signalling/test-signalling-server.js')
-require('./webrtc-star/test-instance.js')
-require('./webrtc-star/test-filter.js')
-require('./webrtc-star/test-listen.js')
-require('./webrtc-star/test-dial.js')
-require('./webrtc-star/test-discovery.js')
-require('./webrtc-star/test-valid-connection.js')
