+++ test/transport/listen.spec.js
@@ -4,7 +4,7 @@
 
 const expect = require('chai').expect
 const multiaddr = require('multiaddr')
-const WebRTCStar = require('../../src/webrtc-star')
+const WebRTCStar = require('../../src')
 
 describe('listen', () => {
   let ws
